# Conflicts:
#	Sources/Internal/Base/Any.h
#	Sources/Internal/Base/Private/Any_impl.h
#	Sources/Internal/Base/Private/AutoStorage_impl.h


Former-commit-id: fcbcbccd647e47f9b1569727b97f6dae1473b2a2

+++ Sources/External/cef/Sources/CEFDavaResourceHandler.cpp
@@ -24,7 +24,7 @@ void CEFDavaResourceHandler::GetResponseHeaders(CefRefPtr<CefResponse> response,
                                                 CefString& redirectUrl)
 {
     FileSystem* fs = FileSystem::Instance();
-    uint32 fileSize = 0;
+    uint64 fileSize = 0;
 
     if (!fs->Exists(davaPath))
     {
